Remove special handling of the pause container
