+++ crates/containerd-shim-wasm/src/sys/unix/mod.rs
@@ -1,5 +1,4 @@
 pub mod container;
 pub mod metrics;
 pub mod networking;
-pub mod signals;
 pub mod stdio;
