--- crates/containerd-shim-wasm/src/sys/windows/signals.rs
@@ -1,2 +1,0 @@
-pub use libc::{SIGINT, SIGTERM};
-pub const SIGKILL: i32 = 9;
